Cleanup Main UI
This patch cleans up the main UI for the watchface by splitting the
date over two lines to remove edge bleed and adding a steps icon above
the steps count to make it even. This also removes device that were
unintentionally supported and adds some of the newer released devices.